Time: 23 ms (57.24%), Space: 17.4 MB (57.22%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -16,7 +16,7 @@ public:
                 int p = j + 1, q = n - 1;
                 while (p < q) {
                     // avoid overflow
-                    long long sum = (long long)nums[i] + nums[j] + nums[p] + nums[q];
+                    long long sum =(long long)nums[i] +(long long) nums[j] +(long long) nums[p] + (long long)nums[q];
 
                     if (sum < target) {
                         p++;
